LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/anonymous-block-box-expected.txt

before
inserted
inline
after

FAIL Anchor selection descent into anonymous block boxes. assert_equals: expected 250 but got 150
